+++ cmd/web/webapp.go
@@ -1,4 +1,4 @@
-package webapp
+package web
 
 import (
 	"fmt"
